# Dailu/src/api/Dailu.Api/OpenApi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Dailu.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/habit-user/integrations/{provider}": {
      "delete": {
        "tags": [
          "HabitUser"
        ],
        "description": "Revokes an integration config for the current authenticated user.",
        "operationId": "DeleteIntegrationConfig",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
...
}